crew/requirements.txt
crewai==0.38.1
crewai-tools==0.8.0
python-dotenv==1.0.0
jira==3.13.0
PyGithub==2.1.1
requests==2.31.0
